refactor: license をsettingsディレクトリに入れた
