+++ app/lib/pages/settings/license_page.dart
